+++ src/collision/TriangleMesh.cpp
@@ -217,7 +217,7 @@ bool TriangleMesh::copyData(const TriangleVertexArray& triangleVertexArray, std:
 void TriangleMesh::removeUnusedVertices(Array<bool>& areUsedVertices) {
 
     // For each vertex of the user mesh
-    for (uint32 i=mVertices.size() - 1; i > 0; i--) {
+    for (uint64 i=mVertices.size() - 1; i > 0; i--) {
 
         // If the vertex is not used
         if (!areUsedVertices[i]) {
@@ -226,7 +226,7 @@ void TriangleMesh::removeUnusedVertices(Array<bool>& areUsedVertices) {
             mVerticesNormals.removeAt(i);
 
             // For each triangle index of the mesh
-            for (uint32 t=0; t < mTriangles.size(); t++) {
+            for (uint64 t=0; t < mTriangles.size(); t++) {
 
                 assert(mTriangles[t] != i);
